+++ app/components/AccountCard/AccountCard.tsx
@@ -9,7 +9,7 @@ import BakerImage from '@resources/svg/baker.svg';
 import ReadonlyImage from '@resources/svg/read-only.svg';
 import LedgerImage from '@resources/svg/ledger.svg';
 import InfoImage from '@resources/svg/info.svg';
-import { displayAsGTU } from '~/utils/gtu';
+import { displayAsCcd } from '~/utils/ccd';
 import { AccountInfo, Account, AccountStatus, ClassName } from '~/utils/types';
 import {
     isInitialAccount,
@@ -72,7 +72,7 @@ function ShieldedBalance({
                     left={rowLeftSide}
                     right={
                         <h3>
-                            {displayAsGTU(shielded)}
+                            {displayAsCcd(shielded)}
                             {hidden}
                         </h3>
                     }
@@ -222,7 +222,7 @@ export function AccountCardView({
                 left={<h3>Account total:</h3>}
                 right={
                     <h3>
-                        {displayAsGTU(shielded + unShielded)}
+                        {displayAsCcd(shielded + unShielded)}
                         {hidden}
                     </h3>
                 }
@@ -231,17 +231,17 @@ export function AccountCardView({
             <SidedRow
                 className={styles.row}
                 left={<h3>Balance:</h3>}
-                right={<h3>{displayAsGTU(unShielded)}</h3>}
+                right={<h3>{displayAsCcd(unShielded)}</h3>}
             />
             <SidedRow
                 className={styles.row}
                 left="- At disposal:"
-                right={displayAsGTU(amountAtDisposal)}
+                right={displayAsCcd(amountAtDisposal)}
             />
             <SidedRow
                 className={styles.row}
                 left="- Staked:"
-                right={displayAsGTU(stakedAmount)}
+                right={displayAsCcd(stakedAmount)}
             />
             <div className={styles.dividingLine} />
             <ShieldedBalance