+++ app/utils/transactionHelpers.ts
@@ -47,7 +47,7 @@ import {
     getUpdateAccountCredentialEnergy,
     getPayloadSizeEstimate,
 } from './transactionCosts';
-import { toMicroUnits, isValidGTUString, displayAsGTU } from './gtu';
+import { ccdToMicroCcd, isValidCcdString, displayAsCcd } from './ccd';
 import { getEncodedSize } from './cborHelper';
 import externalConstants from '~/constants/externalConstants.json';
 import { isASCII } from './basicHelpers';
@@ -520,10 +520,10 @@ export function validateShieldedAmount(
     accountInfo: AccountInfo | undefined,
     estimatedFee: bigint | undefined
 ): string | undefined {
-    if (!isValidGTUString(amountToValidate)) {
+    if (!isValidCcdString(amountToValidate)) {
         return 'Value is not a valid CCD amount';
     }
-    const amountToValidateMicroGTU = toMicroUnits(amountToValidate);
+    const amountToValidateMicroGTU = ccdToMicroCcd(amountToValidate);
     if (
         accountInfo &&
         getAmountAtDisposal(accountInfo) < (estimatedFee || 0n)
@@ -547,10 +547,10 @@ export function validateTransferAmount(
     accountInfo: AccountInfo | undefined,
     estimatedFee: bigint | undefined
 ): string | undefined {
-    if (!isValidGTUString(amountToValidate)) {
+    if (!isValidCcdString(amountToValidate)) {
         return 'Value is not a valid CCD amount';
     }
-    const amountToValidateMicroGTU = toMicroUnits(amountToValidate);
+    const amountToValidateMicroGTU = ccdToMicroCcd(amountToValidate);
     if (
         accountInfo &&
         getAmountAtDisposal(accountInfo) <
@@ -583,12 +583,12 @@ export function validateBakerStake(
     accountInfo: AccountInfo | undefined,
     estimatedFee: bigint | undefined
 ): string | undefined {
-    if (!isValidGTUString(amountToValidate)) {
+    if (!isValidCcdString(amountToValidate)) {
         return 'Value is not a valid CCD amount';
     }
-    const amount = toMicroUnits(amountToValidate);
+    const amount = ccdToMicroCcd(amountToValidate);
     if (bakerStakeThreshold && bakerStakeThreshold > amount) {
-        return `Stake is below the threshold (${displayAsGTU(
+        return `Stake is below the threshold (${displayAsCcd(
             bakerStakeThreshold
         )}) for baking `;
     }